requirements.txt
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
couchbase>=4.2.0
numpy>=1.24.0
tiktoken>=0.5.0
